requirements.txt
streamlit
pandas
pdfplumber
python-docx
fpdf
pillow
scikit-learn
PyPDF2